test: make PHPUnit consistent with other tests

+++ test/phpunit/Migration/MigratorTest.php
@@ -657,7 +657,7 @@ class MigratorTest extends TestCase {
 		}
 
 		if($missingFiles) {
-			$numToRemove = rand(1, $migLength / 10);
+			$numToRemove = rand(1, (int)($migLength / 10));
 			for($i = 0; $i < $numToRemove; $i++) {
 				$keyToRemove = array_rand($fileList);
 				unset($fileList[$keyToRemove]);
